fix(e2e): harden harness boundaries and cancellation
