+++ LgymApi.E2ETests/Harness/PostgreSqlContainerLease.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics;
 using System.Security.Cryptography;
 using DotNet.Testcontainers.Containers;
 using LgymApi.E2ETests.Configuration;
@@ -8,7 +7,6 @@ namespace LgymApi.E2ETests.Harness;
 
 public sealed class PostgreSqlContainerLease : IAsyncDisposable
 {
-    private const string DockerPrerequisiteMessage = "Docker is unavailable for the E2E PostgreSQL lifecycle. Ensure the Docker daemon is running.";
     private const int PostgreSqlPort = 5432;
     private const string PostgreSqlUsername = "postgres";
     private readonly PostgreSqlContainer _container;
@@ -92,135 +90,4 @@ public sealed class PostgreSqlContainerLease : IAsyncDisposable
 
     private static string CreateRandomValue() => RandomNumberGenerator.GetHexString(16).ToLowerInvariant();
 
-    private static class DockerContainerProbe
-    {
-        private static readonly TimeSpan PollInterval = TimeSpan.FromMilliseconds(250);
-
-        public static async Task EnsureAvailableAsync(CancellationToken timeoutToken, CancellationToken callerToken)
-        {
-            using var process = StartDockerVersion();
-            var standardErrorTask = process.StandardError.ReadToEndAsync(timeoutToken);
-            var standardOutputTask = process.StandardOutput.ReadToEndAsync(timeoutToken);
-
-            try
-            {
-                await process.WaitForExitAsync(timeoutToken);
-            }
-            catch (OperationCanceledException) when (!callerToken.IsCancellationRequested)
-            {
-                if (!process.HasExited)
-                {
-                    process.Kill(entireProcessTree: true);
-                }
-
-                throw new InvalidOperationException(DockerPrerequisiteMessage);
-            }
-
-            _ = await standardOutputTask;
-            _ = await standardErrorTask;
-            if (process.ExitCode != 0)
-            {
-                throw new InvalidOperationException(DockerPrerequisiteMessage);
-            }
-        }
-
-        public static async Task<bool> WaitUntilAbsentAsync(string containerId, TimeSpan timeout)
-        {
-            var deadline = DateTimeOffset.UtcNow.Add(timeout);
-
-            while (true)
-            {
-                var remaining = deadline - DateTimeOffset.UtcNow;
-                if (remaining <= TimeSpan.Zero)
-                {
-                    return false;
-                }
-
-                using var inspectionTimeout = new CancellationTokenSource(remaining);
-                if (await IsAbsentAsync(containerId, inspectionTimeout.Token))
-                {
-                    return true;
-                }
-
-                remaining = deadline - DateTimeOffset.UtcNow;
-                if (remaining <= TimeSpan.Zero)
-                {
-                    return false;
-                }
-
-                await Task.Delay(remaining < PollInterval ? remaining : PollInterval);
-            }
-        }
-
-        private static async Task<bool> IsAbsentAsync(string containerId, CancellationToken cancellationToken)
-        {
-            var startInfo = new ProcessStartInfo("docker")
-            {
-                RedirectStandardOutput = true,
-                RedirectStandardError = true
-            };
-            startInfo.ArgumentList.Add("inspect");
-            startInfo.ArgumentList.Add("--type");
-            startInfo.ArgumentList.Add("container");
-            startInfo.ArgumentList.Add("--format");
-            startInfo.ArgumentList.Add("{{.Id}}");
-            startInfo.ArgumentList.Add(containerId);
-
-            using var process = Process.Start(startInfo)
-                ?? throw new InvalidOperationException("Could not start Docker inspection for PostgreSQL cleanup.");
-            var standardErrorTask = process.StandardError.ReadToEndAsync(cancellationToken);
-            var standardOutputTask = process.StandardOutput.ReadToEndAsync(cancellationToken);
-
-            try
-            {
-                await process.WaitForExitAsync(cancellationToken);
-            }
-            catch (OperationCanceledException)
-            {
-                if (!process.HasExited)
-                {
-                    process.Kill(entireProcessTree: true);
-                }
-
-                throw new InvalidOperationException("Docker inspection exceeded the configured shutdown timeout.");
-            }
-
-            _ = await standardOutputTask;
-            var standardError = await standardErrorTask;
-            if (process.ExitCode == 0)
-            {
-                return false;
-            }
-
-            if (standardError.Contains("No such object", StringComparison.OrdinalIgnoreCase) ||
-                standardError.Contains("No such container", StringComparison.OrdinalIgnoreCase))
-            {
-                return true;
-            }
-
-            throw new InvalidOperationException($"Docker inspection failed with exit code {process.ExitCode} during PostgreSQL cleanup.");
-        }
-
-        private static Process StartDockerVersion()
-        {
-            var startInfo = new ProcessStartInfo("docker")
-            {
-                RedirectStandardOutput = true,
-                RedirectStandardError = true
-            };
-            startInfo.ArgumentList.Add("version");
-            startInfo.ArgumentList.Add("--format");
-            startInfo.ArgumentList.Add("{{.Server.Version}}");
-
-            try
-            {
-                return Process.Start(startInfo)
-                    ?? throw new InvalidOperationException(DockerPrerequisiteMessage);
-            }
-            catch (System.ComponentModel.Win32Exception)
-            {
-                throw new InvalidOperationException(DockerPrerequisiteMessage);
-            }
-        }
-    }
 }
